requirements.txt
psycopg2-binary
Pillow
djangorestframework
djangorestframework-simplejwt
pytest
pytest-django
django-cors-headers
drf-yasg==1.21.7
PyYAML==6.0.1
uritemplate==4.1.1
